--- proto/beacon/rpc/v1/BUILD.bazel
@@ -1,61 +1,0 @@
-load("@rules_proto//proto:defs.bzl", "proto_library")
-
-# gazelle:ignore
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
-
-go_proto_library(
-    name = "v1_go_proto",
-    compilers = ["//:cast_grpc_proto_compiler"],
-    importpath = "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1",
-    proto = ":v1_proto",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//proto/beacon/p2p/v1:go_default_library",
-        "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_prysmaticlabs_eth2_types//:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
-        "@go_googleapis//google/api:annotations_go_proto",
-        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
-        "//proto/eth/ext:go_default_library",
-    ],
-)
-go_proto_library(
-    name = "go_grpc_gateway_library",
-    compilers = [
-        "//:grpc_gateway_proto_compiler",
-    ],
-    embed = ["v1_go_proto"],
-    importpath = "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1",
-    proto = ":v1_proto",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//proto/beacon/p2p/v1:go_default_library",
-        "@com_github_prysmaticlabs_eth2_types//:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
-        "@go_googleapis//google/api:annotations_go_proto",
-        "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
-        "@io_bazel_rules_go//proto/wkt:empty_go_proto",
-        "//proto/eth/ext:go_default_library",
-    ],
-)
-
-go_library(
-    name = "go_default_library",
-    embed = [":go_grpc_gateway_library"],
-    importpath = "github.com/prysmaticlabs/prysm/proto/beacon/rpc/v1",
-    visibility = ["//visibility:public"],
-)
-
-proto_library(
-    name = "v1_proto",
-    srcs = ["debug.proto", "health.proto", "slasher.proto"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//proto/beacon/p2p/v1:v1_proto",
-        "//proto/prysm/v1alpha1:proto",
-        "//proto/eth/ext:proto",
-        "@com_google_protobuf//:empty_proto",
-        "@go_googleapis//google/api:annotations_proto",
-    ],
-)
