+++ proto/prysm/v2/slasher.proto
@@ -1,12 +1,19 @@
 syntax = "proto3";
 
-package ethereum.beacon.rpc.v1;
+package ethereum.prysm.v2;
 
 import "proto/eth/ext/options.proto";
 import "proto/prysm/v1alpha1/beacon_block.proto";
 
 import "google/api/annotations.proto";
 import "google/protobuf/empty.proto";
+
+option csharp_namespace = "Ethereum.Prysm.V2";
+option go_package = "github.com/prysmaticlabs/prysm/proto/prysm/v2;v2";
+option java_multiple_files = true;
+option java_outer_classname = "SlasherProto";
+option java_package = "org.ethereum.prysm.v2";
+option php_namespace = "Ethereum\\Prysm\\v2";
 
 // Slasher service API
 //
